requirements.txt
airtable-python-wrapper==0.11.3.post1
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
idna==2.8
passlib==1.7.1
pycparser==2.19
requests==2.21.0
six==1.12.0
urllib3==1.24.1
